requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
argilla==1.12.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.3.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
certifi==2023.5.7
cffi==1.15.1
chainlit==0.5.2
chardet==5.1.0
charset-normalizer==3.2.0
chromadb==0.3.23
click==8.1.6
clickhouse-connect==0.6.8
colorclass==2.2.2
commonmark==0.9.1
compressed-rtf==1.0.6
cryptography==41.0.2
dataclasses-json==0.5.12
Deprecated==1.2.14
duckdb==0.8.1
easygui==0.98.3
ebcdic==1.1.1
et-xmlfile==1.1.0
extract-msg==0.41.1
faiss-cpu==1.7.4
fastapi==0.97.0
fastapi-socketio==0.0.10
filelock==3.12.2
filetype==1.2.0
frozenlist==1.4.0
fsspec==2023.6.0
googleapis-common-protos==1.59.1
gpt4all==0.3.4
grpcio==1.56.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
huggingface-hub==0.16.4
idna==3.4
IMAPClient==2.3.1
importlib-metadata==6.0.1
Jinja2==3.1.2
joblib==1.3.1
langchain==0.0.238
langchainplus-sdk==0.0.20
langsmith==0.0.11
lark-parser==0.12.0
llama-cpp-python==0.1.50
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msoffcrypto-tool==5.1.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
oletools==0.60.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
packaging==23.1
pandas==1.5.3
pandoc==2.3
pcodedmp==1.2.6
pdfminer.six==20221105
Pillow==10.0.0
plumbum==1.8.2
ply==3.11
posthog==3.0.1
prisma==0.9.1
protobuf==4.23.4
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.8.0
PyMuPDF==1.22.3
pypandoc==1.11
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.5.1
python-graphql-client==0.4.3
python-magic==0.4.27
python-pptx==0.6.21
python-socketio==5.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
red-black-tree-mod==1.20
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
RTFDE==0.0.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
starlette==0.27.0
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.2
unstructured==0.6.6
uptrace==1.18.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.14.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
